Rename directory to pass uBlock Origin
