+++ web-common/src/features/metrics-layer/workspace/MetricsGenerateButton.svelte
