+++ web-common/src/features/metrics-layer/workspace/MetricsLayoutManager.svelte
