+++ web-common/src/features/metrics-layer/workspace/QuickMetricsModal.svelte
