+++ web-local/src/lib/components/table-editable/ColumnConfig.ts
@@ -1,6 +1,6 @@
 import type { EntityRecord } from "@rilldata/web-common/features/entity-management/types";
 import type { SvelteComponent } from "svelte";
-import type { ValidationState } from "../../../../../web-common/src/features/metrics/errors";
+import type { ValidationState } from "../../../../../web-common/src/features/metrics-layer/errors";
 import TableCellInput from "./TableCellInput.svelte";
 import TableCellSelector from "./TableCellSelector.svelte";
